+++ nodes/decoder.js
@@ -561,6 +561,22 @@ module.exports = {
       deviceId: uint32(bytes, 4),
       index: uint32(bytes, 8)
     }
+  },
+
+  /**
+    * Decodes the response to a set-pc-control request (function code 0xa0).
+    *
+    * @param {buffer}   buffer     64 byte NodeJS buffer
+    * @param {function} translator (optional) function to internationalise the text in a
+    *                              decoded object
+    *
+    * @param {object}   Decoded set-pc-control response object
+    */
+  SetPCControl: function (bytes, translator) {
+    return {
+      deviceId: uint32(bytes, 4),
+      updated: bool(bytes, 8)
+    }
   }
 }