+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## [Unreleased]
 
+### Added
+1. uhppoted-set-pc-control node.
+
+### Updated
+1. Changed package.json to support NodeRED v2.2.3+
+
 
 ## [1.1.2](https://github.com/uhppoted/node-red-contrib-uhppoted/releases/tag/v1.1.2) - 2022-12-16